chore(ipld): extract byzantine error and constructor for it into a separate file
