+++ ipld/retriever.go
@@ -14,7 +14,6 @@ import (
 	"github.com/tendermint/tendermint/pkg/da"
 	"github.com/tendermint/tendermint/pkg/wrapper"
 
-	"github.com/celestiaorg/celestia-node/ipld/plugin"
 	"github.com/celestiaorg/nmt"
 	"github.com/celestiaorg/rsmt2d"
 )
@@ -58,8 +57,12 @@ func (r *Retriever) Retrieve(ctx context.Context, dah *da.DataAvailabilityHeader
 			if err == nil {
 				return eds, nil
 			}
-			if byzErr := r.checkForByzantineError(ctx, dah, err); byzErr != nil {
-				return nil, byzErr
+			var (
+				errRow *rsmt2d.ErrByzantineRow
+				errCol *rsmt2d.ErrByzantineCol
+			)
+			if errors.As(err, &errRow) || errors.As(err, &errCol) {
+				return nil, NewErrByzantine(ctx, r.dag, dah, errRow, errCol)
 			}
 			log.Warnw("not enough shares to reconstruct data square, requesting more...", "err", err)
 		}
@@ -67,50 +70,6 @@ func (r *Retriever) Retrieve(ctx context.Context, dah *da.DataAvailabilityHeader
 	}
 
 	return nil, format.ErrNotFound
-}
-
-// checkForByzantineError ensures that passed error is rsmt2d.ErrByzantineRow/Col,
-// fetches proof for every share from row or col and converts error to ErrByzantine
-func (r *Retriever) checkForByzantineError(
-	ctx context.Context,
-	dah *da.DataAvailabilityHeader,
-	byzErr error,
-) error {
-	var (
-		errRow *rsmt2d.ErrByzantineRow
-		errCol *rsmt2d.ErrByzantineCol
-	)
-	if !errors.As(byzErr, &errRow) && !errors.As(byzErr, &errCol) {
-		return nil
-	}
-	var (
-		errShares [][]byte
-		root      []byte
-		index     uint8
-	)
-	isRow := false
-	if errRow != nil {
-		errShares = errRow.Shares
-		root = dah.RowsRoots[errRow.RowNumber]
-		index = uint8(errRow.RowNumber)
-		isRow = true
-	} else {
-		errShares = errCol.Shares
-		root = dah.ColumnRoots[errCol.ColNumber]
-		index = uint8(errCol.ColNumber)
-	}
-
-	sharesWithProof, err := GetProofsForShares(
-		ctx,
-		r.dag,
-		plugin.MustCidFromNamespacedSha256(root),
-		errShares,
-	)
-	if err != nil {
-		return err
-	}
-
-	return &ErrByzantine{Index: index, Shares: sharesWithProof, IsRow: isRow}
 }
 
 type retrieverSession struct {
